Adds -Wextra
Use -Wextra and fixes many implicit casts and unused variables.

+++ libchronos/chronos-physical.c
@@ -184,8 +184,8 @@ int chronos_entry( struct chronos_handle * handle, int entry_number, struct inde
 
 	struct index_entry platform_agnostic;
 
-	int read_data = read( handle->index_fd, & platform_agnostic, sizeof(platform_agnostic) );
-	if( read_data == -1 || read_data < sizeof(platform_agnostic) ){
+	ssize_t read_data = read( handle->index_fd, & platform_agnostic, sizeof(platform_agnostic) );
+	if( read_data == -1 || read_data < (ssize_t)sizeof(platform_agnostic) ){
 		return C_IO_READ_ERROR;
 	}
